// src/utils/models/Neighbour.model.ts
export class NeighbourModel{
    _id: string;
    id_usuario: string;
    num_inmueble: string;
    id_establecimiento: string;
    email: string;
    estado: string;
    nombres: string;
    numero_celular: string;
    apellidos: string;
    tipo_doc : string;
    identificacion: string;
    sexo: string;
    fecha_nacimiento: string;
    foto: string;
}